cloud_functions/load-v2-tournament-data-3/requirements.txt
numerapi==2.8.1
google-cloud-bigquery==1.21.0
google-cloud-pubsub==2.8.0
